Apply GlobalSnapshotManager optimization from upstream
The same as in https://github.com/androidx/androidx/commit/f468727e328aad5c5baef9b5b3548c53e19fed2d

This optimization is valuable for some cases. it isn't visible in our current benchmarks, but was visible in some androidx benchmarks for Android (LazyList scrolling)
