+++ compose/ui/ui/src/darwinMain/kotlin/androidx/compose/ui/platform/GlobalSnapshotManager.native.kt
@@ -16,39 +16,7 @@
 
 package androidx.compose.ui.platform
 
-import androidx.compose.runtime.snapshots.Snapshot
-import androidx.compose.ui.platform.GlobalSnapshotManager.ensureStarted
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.channels.Channel
-import kotlinx.coroutines.channels.consumeEach
-import kotlinx.coroutines.launch
+import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Dispatchers
 
-/**
- * Platform-specific mechanism for starting a monitor of global snapshot state writes
- * in order to schedule the periodic dispatch of snapshot apply notifications.
- * This process should remain platform-specific; it is tied to the threading and update model of
- * a particular platform and framework target.
- *
- * Composition bootstrapping mechanisms for a particular platform/framework should call
- * [ensureStarted] during setup to initialize periodic global snapshot notifications.
- * For UIKit, these notifications are always sent on [Dispatchers.Main]. Other platforms
- * may establish different policies for these notifications.
- */
-internal actual object GlobalSnapshotManager {
-    private val started = AtomicInt(0)
-
-    actual fun ensureStarted() {
-        if (started.compareAndSet(0, 1)) {
-            val channel = Channel<Unit>(Channel.CONFLATED)
-            CoroutineScope(Dispatchers.Main).launch {
-                channel.consumeEach {
-                    Snapshot.sendApplyNotifications()
-                }
-            }
-            Snapshot.registerGlobalWriteObserver {
-                channel.trySend(Unit)
-            }
-        }
-    }
-}
+internal actual val GlobalSnapshotManagerDispatcher: CoroutineDispatcher = Dispatchers.Main
