+++ compose/ui/ui/src/skikoMain/kotlin/androidx/compose/ui/platform/GlobalSnapshotManager.skiko.kt
@@ -16,6 +16,13 @@
 
 package androidx.compose.ui.platform
 
+import androidx.compose.runtime.snapshots.Snapshot
+import kotlinx.coroutines.CoroutineDispatcher
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.channels.Channel
+import kotlinx.coroutines.channels.consumeEach
+import kotlinx.coroutines.launch
+
 /**
  * Platform-specific mechanism for starting a monitor of global snapshot state writes
  * in order to schedule the periodic dispatch of snapshot apply notifications.
@@ -25,6 +32,26 @@ package androidx.compose.ui.platform
  * Composition bootstrapping mechanisms for a particular platform/framework should call
  * [ensureStarted] during setup to initialize periodic global snapshot notifications.
  */
-internal expect object GlobalSnapshotManager {
-    fun ensureStarted()
+internal object GlobalSnapshotManager {
+    private val started = AtomicInt(0)
+    private val sent = AtomicInt(0)
+
+    fun ensureStarted() {
+        if (started.compareAndSet(0, 1)) {
+            val channel = Channel<Unit>(1)
+            CoroutineScope(GlobalSnapshotManagerDispatcher).launch {
+                channel.consumeEach {
+                    sent.compareAndSet(1, 0)
+                    Snapshot.sendApplyNotifications()
+                }
+            }
+            Snapshot.registerGlobalWriteObserver {
+                if (sent.compareAndSet(0, 1)) {
+                    channel.trySend(Unit)
+                }
+            }
+        }
+    }
 }
+
+internal expect val GlobalSnapshotManagerDispatcher: CoroutineDispatcher
